client: Split out user and team apis
Just some house keeping we've been gradually doing over time.
